backend/requirements.txt
# Core ML & Computer Vision (Lightweight)
opencv-python-headless>=4.5.0,<5.0.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
pandas>=1.5.0,<3.0.0

# Image & PDF Processing
Pillow>=9.0.0,<11.0.0
pdf2image>=1.16.0
pdfplumber>=0.7.0

# OCR Engine (Lightweight - Tesseract only)
pytesseract>=0.3.8

# Web Framework
fastapi>=0.95.0,<1.0.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
aiofiles>=22.0.0

# Rate Limiting
slowapi>=0.1.8

# Database
sqlalchemy>=2.0.0,<3.0.0

# Authentication & Security
pyjwt>=2.6.0
bcrypt>=4.0.0
python-dotenv>=1.0.0

# Email
email-validator>=1.3.0

# Date/Time Utilities
python-dateutil>=2.8.0

# Data Validation
pydantic>=1.10.0,<2.0.0

# Logging
python-json-logger>=2.0.0

# System Monitoring (for health checks)
psutil>=5.9.0

# HTTP Client (for health checks)
httpx>=0.24.0

# Build Tools (Python 3.11 compatibility)
setuptools>=65.0.0
wheel>=0.38.0

# Additional dependencies for Hugging Face Spaces
gunicorn>=20.1.0